requirements.txt
requests==2.19.1
PyYAML==6.0